fix: refactor errorNotBusinessUnit
