--- src/hooks/useClearLocalStorageOnMount.ts
@@ -1,9 +1,0 @@
-import { useEffect } from "react";
-
-const useClearLocalStorageOnMount = () => {
-  useEffect(() => {
-    localStorage.clear();
-  }, []);
-};
-
-export { useClearLocalStorageOnMount };
